corrected optional import in cfdmod utils
